requirements.txt
pyzmq>=25.0.0
opencv-python>=4.8.0
pyserial>=3.5
flask>=3.0.0
flask-socketio>=5.3.0
ultralytics>=8.3.0
numpy>=1.24.0
filterpy>=1.4.5

# 语音交互依赖
sherpa-onnx>=1.9.0
sounddevice>=0.4.6
openai>=1.0.0
fastmcp>=2.14.0
websockets>=12.0

# 火山引擎LLM依赖
volcengine-python-sdk
